main.c: Add host_access mode.
Add a mode whereby the host CPU accesses the p2pmem device directly
using load/store accesses on the p2pmem device. If requested do both
writes and reads and do a check to ensure the data read agrees with
what was written.

When --host_access is 0 no test is performed. When >0 then that value
is used as the alignment and size of the access for both reads and
writes. When <0 use -1*value for alignment and size but only do a read
test. Use --check to perform a check on data (only when --host_access
> 0).