main.c: Add support for multi-threading
Add a --threads options that creates this many worker threads. Each
thread operates on its own contigious section of the read and write
devices. Each thread claims a dsicreate, contigious section of the
p2pmem device or system memory buffer. We perform some checks on the
input arguments to make sure the user requests sane things.

Note that the size of the p2pmem device is checked via the mmap
call. It will fail if you ask for more memory than the device can
provide.