README.md: Updates to README.md for Linux kernel v6.2
In v6.2 Logan added code that now exposes NVMe CMBs to userspace via a
sysfs entry. There are no changes to the p2pmem-test source code but
we do want to update the README.md to reflect that OOT modules are no
longer needed. While we are at it we do a bit of a clean-up of
README.md.xsx

Fixes #11.